Added sketch folder.
