+++ Igra/hal_conf_extra.h
