+++ Igra/myDefines.h
