--- lab.h
@@ -1,26 +1,0 @@
-
-#ifndef __LAB_H__ 
-#define __LAB_H__ 
-#include "myDefines.h" 
-#include "ILI9341_STM32.h" 
-const myline_t lab1Lines[] = {{3, 5, 128, 5}, {159, 5, 315, 5}, {3, 46, 35, 46}, {66, 46, 97, 46}, {159, 46, 190, 46}, {221, 46, 252, 46}, {35, 88, 66, 88}, {97, 88, 159, 88}, {221, 88, 283, 88}, {66, 130, 97, 130}, {159, 130, 221, 130}, {252, 130, 283, 130}, {35, 172, 66, 172}, {97, 172, 190, 172}, {221, 172, 252, 172}, {3, 213, 159, 213}, {190, 213, 315, 213}, {3, 5, 3, 213}, {35, 88, 35, 130}, {66, 46, 66, 88}, {66, 130, 66, 213}, {97, 46, 97, 130}, {128, 5, 128, 46}, {128, 130, 128, 172}, {159, 46, 159, 130}, {190, 46, 190, 88}, {190, 172, 190, 213}, {221, 46, 221, 88}, {221, 130, 221, 172}, {252, 130, 252, 172}, {283, 5, 283, 88}, {283, 172, 283, 213}, {315, 5, 315, 213}};
-const myMaze_t lab1 = {(myline_t*)lab1Lines ,143 , 5  , {166, 213, 184, 213 },sizeof(lab1Lines) / sizeof(myline_t)};
-const myline_t lab2Lines[] = {{3, 5, 128, 5}, {159, 5, 315, 5}, {97, 46, 221, 46}, {252, 46, 283, 46}, {35, 88, 128, 88}, {159, 88, 252, 88}, {128, 130, 190, 130}, {97, 172, 128, 172}, {190, 172, 283, 172}, {3, 213, 159, 213}, {190, 213, 315, 213}, {3, 5, 3, 213}, {35, 46, 35, 88}, {35, 130, 35, 213}, {66, 5, 66, 46}, {66, 88, 66, 172}, {97, 130, 97, 213}, {128, 46, 128, 88}, {159, 130, 159, 213}, {190, 88, 190, 130}, {190, 172, 190, 213}, {221, 46, 221, 88}, {221, 130, 221, 172}, {252, 88, 252, 130}, {283, 46, 283, 172}, {315, 5, 315, 213}};
-const myMaze_t lab2 = {(myline_t*)lab2Lines ,143 , 5  , {166, 213, 184, 213 },sizeof(lab2Lines) / sizeof(myline_t)};
-const myline_t lab3Lines[] = {{3, 4, 145, 4}, {173, 4, 315, 4}, {31, 39, 88, 39}, {173, 39, 202, 39}, {116, 74, 145, 74}, {202, 74, 287, 74}, {31, 109, 116, 109}, {145, 109, 258, 109}, {287, 109, 315, 109}, {3, 144, 88, 144}, {173, 144, 202, 144}, {88, 179, 173, 179}, {202, 179, 230, 179}, {3, 214, 145, 214}, {173, 214, 315, 214}, {3, 4, 3, 214}, {31, 39, 31, 109}, {31, 179, 31, 214}, {60, 74, 60, 109}, {60, 144, 60, 179}, {88, 39, 88, 74}, {116, 4, 116, 74}, {116, 109, 116, 179}, {145, 39, 145, 74}, {145, 109, 145, 144}, {173, 39, 173, 109}, {173, 144, 173, 214}, {230, 4, 230, 74}, {230, 109, 230, 179}, {258, 39, 258, 74}, {258, 109, 258, 214}, {287, 4, 287, 39}, {287, 109, 287, 179}, {315, 4, 315, 214}};
-const myMaze_t lab3 = {(myline_t*)lab3Lines ,159 , 4  , {150, 214, 169, 214 },sizeof(lab3Lines) / sizeof(myline_t)};
-const myline_t lab4Lines[] = {{3, 4, 145, 4}, {173, 4, 315, 4}, {31, 39, 60, 39}, {88, 39, 116, 39}, {258, 39, 287, 39}, {31, 74, 60, 74}, {88, 74, 202, 74}, {230, 74, 258, 74}, {3, 109, 31, 109}, {60, 109, 88, 109}, {116, 109, 230, 109}, {258, 109, 287, 109}, {230, 144, 258, 144}, {60, 179, 116, 179}, {202, 179, 230, 179}, {3, 214, 145, 214}, {173, 214, 315, 214}, {3, 4, 3, 214}, {31, 109, 31, 179}, {60, 4, 60, 39}, {60, 74, 60, 214}, {88, 109, 88, 144}, {116, 39, 116, 74}, {116, 144, 116, 179}, {145, 4, 145, 39}, {145, 109, 145, 214}, {173, 39, 173, 74}, {173, 144, 173, 214}, {202, 4, 202, 74}, {202, 109, 202, 179}, {230, 39, 230, 109}, {258, 39, 258, 74}, {258, 109, 258, 214}, {287, 74, 287, 179}, {315, 4, 315, 214}};
-const myMaze_t lab4 = {(myline_t*)lab4Lines ,159 , 4  , {150, 214, 169, 214 },sizeof(lab4Lines) / sizeof(myline_t)};
-const myline_t lab5Lines[] = {{3, 3, 133, 3}, {159, 3, 315, 3}, {237, 33, 263, 33}, {3, 64, 29, 64}, {55, 64, 211, 64}, {237, 64, 289, 64}, {29, 94, 55, 94}, {211, 94, 237, 94}, {263, 94, 315, 94}, {3, 124, 29, 124}, {107, 124, 133, 124}, {237, 124, 263, 124}, {29, 154, 185, 154}, {263, 154, 289, 154}, {81, 185, 211, 185}, {263, 185, 315, 185}, {3, 215, 159, 215}, {185, 215, 315, 215}, {3, 3, 3, 215}, {29, 33, 29, 64}, {29, 154, 29, 185}, {55, 33, 55, 124}, {55, 154, 55, 215}, {81, 3, 81, 33}, {81, 94, 81, 154}, {107, 33, 107, 94}, {133, 3, 133, 33}, {133, 64, 133, 124}, {159, 33, 159, 64}, {159, 94, 159, 154}, {185, 3, 185, 33}, {185, 64, 185, 154}, {185, 185, 185, 215}, {211, 33, 211, 64}, {211, 94, 211, 185}, {237, 3, 237, 33}, {237, 64, 237, 94}, {237, 154, 237, 215}, {263, 94, 263, 154}, {289, 33, 289, 64}, {289, 124, 289, 154}, {315, 3, 315, 215}};
-const myMaze_t lab5 = {(myline_t*)lab5Lines ,146 , 3  , {164, 215, 181, 215 },sizeof(lab5Lines) / sizeof(myline_t)};
-const myline_t lab6Lines[] = {{3, 3, 147, 3}, {171, 3, 315, 3}, {3, 29, 75, 29}, {99, 29, 123, 29}, {147, 29, 195, 29}, {123, 56, 147, 56}, {195, 56, 219, 56}, {267, 56, 315, 56}, {27, 82, 123, 82}, {195, 82, 243, 82}, {267, 82, 291, 82}, {51, 109, 75, 109}, {171, 109, 219, 109}, {243, 109, 291, 109}, {27, 136, 51, 136}, {75, 136, 99, 136}, {123, 136, 171, 136}, {195, 136, 243, 136}, {291, 136, 315, 136}, {51, 162, 75, 162}, {99, 162, 195, 162}, {219, 162, 291, 162}, {27, 189, 51, 189}, {75, 189, 99, 189}, {171, 189, 219, 189}, {243, 189, 315, 189}, {3, 215, 147, 215}, {171, 215, 315, 215}, {3, 3, 3, 215}, {27, 56, 27, 189}, {51, 29, 51, 56}, {75, 56, 75, 82}, {75, 109, 75, 136}, {75, 162, 75, 215}, {99, 29, 99, 56}, {99, 82, 99, 162}, {123, 3, 123, 29}, {123, 109, 123, 136}, {123, 162, 123, 189}, {147, 29, 147, 109}, {147, 162, 147, 215}, {171, 56, 171, 136}, {171, 189, 171, 215}, {195, 29, 195, 82}, {195, 136, 195, 162}, {219, 3, 219, 29}, {219, 109, 219, 136}, {219, 162, 219, 189}, {243, 29, 243, 109}, {267, 3, 267, 29}, {267, 109, 267, 162}, {291, 29, 291, 56}, {291, 82, 291, 109}, {315, 3, 315, 215}};
-const myMaze_t lab6 = {(myline_t*)lab6Lines ,159 , 3  , {151, 215, 168, 215 },sizeof(lab6Lines) / sizeof(myline_t)};
-const myline_t lab7Lines[] = {{3, 3, 147, 3}, {171, 3, 315, 3}, {27, 29, 51, 29}, {99, 29, 171, 29}, {3, 56, 27, 56}, {51, 56, 75, 56}, {171, 56, 267, 56}, {291, 82, 315, 82}, {51, 109, 123, 109}, {171, 109, 219, 109}, {243, 109, 315, 109}, {27, 136, 99, 136}, {147, 136, 195, 136}, {243, 136, 291, 136}, {51, 162, 75, 162}, {123, 162, 171, 162}, {267, 162, 315, 162}, {27, 189, 51, 189}, {99, 189, 171, 189}, {219, 189, 267, 189}, {3, 215, 147, 215}, {171, 215, 315, 215}, {3, 3, 3, 215}, {27, 82, 27, 189}, {51, 29, 51, 109}, {75, 3, 75, 29}, {75, 56, 75, 82}, {75, 162, 75, 215}, {99, 29, 99, 109}, {99, 136, 99, 189}, {123, 56, 123, 162}, {147, 29, 147, 136}, {171, 29, 171, 82}, {171, 189, 171, 215}, {195, 3, 195, 29}, {195, 82, 195, 109}, {195, 136, 195, 215}, {219, 29, 219, 162}, {243, 3, 243, 29}, {243, 82, 243, 136}, {243, 162, 243, 189}, {267, 29, 267, 82}, {267, 162, 267, 189}, {291, 3, 291, 56}, {291, 189, 291, 215}, {315, 3, 315, 215}};
-const myMaze_t lab7 = {(myline_t*)lab7Lines ,159 , 3  , {151, 215, 168, 215 },sizeof(lab7Lines) / sizeof(myline_t)};
-const myline_t lab8Lines[] = {{2, 2, 137, 2}, {159, 2, 316, 2}, {47, 26, 92, 26}, {114, 26, 137, 26}, {159, 26, 204, 26}, {271, 26, 293, 26}, {2, 50, 47, 50}, {92, 50, 114, 50}, {137, 50, 159, 50}, {249, 50, 271, 50}, {25, 73, 69, 73}, {92, 73, 181, 73}, {226, 73, 249, 73}, {271, 73, 293, 73}, {2, 97, 47, 97}, {69, 97, 92, 97}, {114, 97, 159, 97}, {204, 97, 226, 97}, {249, 97, 271, 97}, {293, 97, 316, 97}, {25, 121, 69, 121}, {159, 121, 181, 121}, {204, 121, 249, 121}, {271, 121, 293, 121}, {25, 145, 69, 145}, {92, 145, 137, 145}, {159, 145, 204, 145}, {226, 145, 271, 145}, {2, 168, 25, 168}, {69, 168, 114, 168}, {137, 168, 159, 168}, {204, 168, 249, 168}, {271, 168, 293, 168}, {47, 192, 69, 192}, {114, 192, 137, 192}, {159, 192, 181, 192}, {293, 192, 316, 192}, {2, 216, 159, 216}, {181, 216, 316, 216}, {2, 2, 2, 216}, {25, 2, 25, 26}, {25, 121, 25, 145}, {25, 192, 25, 216}, {47, 145, 47, 192}, {69, 26, 69, 73}, {69, 97, 69, 121}, {92, 2, 92, 26}, {92, 50, 92, 145}, {92, 168, 92, 216}, {114, 26, 114, 50}, {114, 97, 114, 121}, {137, 121, 137, 216}, {159, 26, 159, 50}, {159, 97, 159, 121}, {159, 145, 159, 168}, {181, 50, 181, 121}, {181, 168, 181, 216}, {204, 26, 204, 97}, {204, 121, 204, 192}, {226, 26, 226, 73}, {226, 97, 226, 121}, {226, 192, 226, 216}, {249, 2, 249, 50}, {249, 73, 249, 97}, {249, 145, 249, 192}, {271, 73, 271, 121}, {271, 168, 271, 216}, {293, 26, 293, 73}, {293, 121, 293, 168}, {316, 2, 316, 216}};
-const myMaze_t lab8 = {(myline_t*)lab8Lines ,148 , 2  , {162, 216, 179, 216 },sizeof(lab8Lines) / sizeof(myline_t)};
-const myline_t lab9Lines[] = {{2, 2, 139, 2}, {159, 2, 316, 2}, {41, 21, 80, 21}, {100, 21, 120, 21}, {139, 21, 159, 21}, {179, 21, 198, 21}, {218, 21, 257, 21}, {22, 41, 61, 41}, {100, 41, 198, 41}, {218, 41, 238, 41}, {61, 60, 100, 60}, {120, 60, 159, 60}, {179, 60, 218, 60}, {238, 60, 296, 60}, {22, 80, 61, 80}, {100, 80, 120, 80}, {159, 80, 198, 80}, {257, 80, 316, 80}, {2, 99, 22, 99}, {61, 99, 80, 99}, {120, 99, 139, 99}, {179, 99, 198, 99}, {238, 99, 296, 99}, {22, 119, 100, 119}, {139, 119, 179, 119}, {218, 119, 257, 119}, {277, 119, 316, 119}, {61, 138, 80, 138}, {120, 138, 198, 138}, {218, 138, 238, 138}, {41, 158, 61, 158}, {80, 158, 139, 158}, {179, 158, 218, 158}, {238, 158, 296, 158}, {61, 177, 80, 177}, {100, 177, 120, 177}, {139, 177, 159, 177}, {218, 177, 238, 177}, {277, 177, 296, 177}, {2, 197, 22, 197}, {61, 197, 80, 197}, {120, 197, 139, 197}, {198, 197, 296, 197}, {2, 216, 159, 216}, {179, 216, 316, 216}, {2, 2, 2, 216}, {22, 21, 22, 41}, {22, 60, 22, 80}, {22, 119, 22, 177}, {41, 2, 41, 21}, {41, 60, 41, 119}, {41, 138, 41, 216}, {61, 41, 61, 60}, {61, 119, 61, 138}, {61, 158, 61, 177}, {80, 21, 80, 41}, {80, 60, 80, 99}, {80, 138, 80, 158}, {80, 177, 80, 197}, {100, 41, 100, 60}, {100, 80, 100, 138}, {100, 197, 100, 216}, {120, 2, 120, 21}, {120, 60, 120, 80}, {120, 99, 120, 138}, {120, 158, 120, 197}, {139, 21, 139, 41}, {139, 80, 139, 99}, {159, 60, 159, 119}, {159, 138, 159, 216}, {179, 2, 179, 21}, {179, 41, 179, 60}, {179, 158, 179, 216}, {198, 80, 198, 138}, {198, 158, 198, 197}, {218, 21, 218, 158}, {238, 60, 238, 99}, {238, 158, 238, 177}, {257, 21, 257, 60}, {257, 119, 257, 158}, {257, 177, 257, 197}, {277, 2, 277, 41}, {277, 119, 277, 138}, {277, 158, 277, 177}, {296, 21, 296, 60}, {296, 138, 296, 158}, {316, 2, 316, 216}};
-const myMaze_t lab9 = {(myline_t*)lab9Lines ,149 , 2  , {161, 216, 177, 216 },sizeof(lab9Lines) / sizeof(myline_t)};
-const myline_t lab10Lines[] = {{2, 2, 139, 2}, {159, 2, 316, 2}, {22, 21, 41, 21}, {61, 21, 100, 21}, {120, 21, 139, 21}, {257, 21, 296, 21}, {41, 41, 61, 41}, {80, 41, 120, 41}, {139, 41, 179, 41}, {218, 41, 257, 41}, {277, 41, 316, 41}, {22, 60, 41, 60}, {61, 60, 80, 60}, {120, 60, 139, 60}, {179, 60, 198, 60}, {218, 60, 238, 60}, {257, 60, 277, 60}, {2, 80, 22, 80}, {41, 80, 61, 80}, {80, 80, 100, 80}, {120, 80, 179, 80}, {198, 80, 218, 80}, {238, 80, 257, 80}, {22, 99, 61, 99}, {100, 99, 120, 99}, {179, 99, 218, 99}, {257, 99, 277, 99}, {296, 99, 316, 99}, {2, 119, 41, 119}, {61, 119, 80, 119}, {120, 119, 139, 119}, {159, 119, 179, 119}, {238, 119, 257, 119}, {2, 138, 22, 138}, {41, 138, 61, 138}, {120, 138, 159, 138}, {179, 138, 238, 138}, {257, 138, 277, 138}, {296, 138, 316, 138}, {22, 158, 61, 158}, {159, 158, 179, 158}, {198, 158, 257, 158}, {2, 177, 80, 177}, {100, 177, 139, 177}, {179, 177, 238, 177}, {277, 177, 316, 177}, {22, 197, 61, 197}, {80, 197, 100, 197}, {139, 197, 198, 197}, {2, 216, 159, 216}, {179, 216, 316, 216}, {2, 2, 2, 216}, {22, 41, 22, 80}, {41, 2, 41, 41}, {41, 80, 41, 99}, {41, 138, 41, 158}, {61, 41, 61, 80}, {61, 119, 61, 138}, {61, 197, 61, 216}, {80, 21, 80, 41}, {80, 80, 80, 197}, {100, 60, 100, 80}, {100, 99, 100, 177}, {120, 2, 120, 60}, {120, 80, 120, 99}, {120, 119, 120, 138}, {120, 158, 120, 177}, {120, 197, 120, 216}, {139, 99, 139, 119}, {139, 138, 139, 158}, {159, 2, 159, 21}, {159, 60, 159, 138}, {159, 158, 159, 216}, {179, 21, 179, 80}, {179, 138, 179, 158}, {198, 2, 198, 60}, {198, 80, 198, 119}, {198, 158, 198, 177}, {218, 21, 218, 80}, {218, 119, 218, 138}, {218, 177, 218, 216}, {238, 2, 238, 21}, {238, 80, 238, 138}, {238, 197, 238, 216}, {257, 21, 257, 80}, {257, 138, 257, 197}, {277, 80, 277, 138}, {277, 158, 277, 177}, {277, 197, 277, 216}, {296, 41, 296, 80}, {296, 99, 296, 119}, {296, 138, 296, 158}, {296, 177, 296, 197}, {316, 2, 316, 216}};
-const myMaze_t lab10 = {(myline_t*)lab10Lines ,149 , 2  , {161, 216, 177, 216 },sizeof(lab10Lines) / sizeof(myline_t)};
-#endif
